DDCNL-9071: Removal of TaxAccountSummaryHodFormatters
